+++ modules/nodes/00-variables_required_inputs.tf
@@ -14,10 +14,19 @@ variable "k8s_workers_num_nodes" {
 }
 
 variable "nodes_cri_bootstrap" {
-  type = string
+  type        = string
+  description = "The CRI to use for this node."
 }
 
-variable "private_subnets" {}
+variable "private_subnets" {
+  type        = list(string)
+  description = "The list of all possible subnets IDS controllers and nodes may be created into"
+}
+
+variable "private_subnets_cidr" {
+  type        = list(string)
+  description = "The list of all possible subnets CIDR controllers and nodes may be created into"
+}
 
 variable "nodes_config_bundle" {
   type = object({
@@ -29,4 +38,5 @@ variable "nodes_config_bundle" {
     secret_arn              = string
     lb_dns                  = string
   })
+  description = "Map which containes all the data requires to spin up and attach a node to a set of controllers. It is an output of the controller module."
 }